Extracted function CheckValidEmail() into an own file
