+++ index.php
@@ -2,6 +2,8 @@
 // Composer autoloading
 require __DIR__ . '/vendor/autoload.php';
 
+require_once('inc/Functions/CheckValidEmail.php');
+
 // Set error_reporting.
 // It is set to this value on purpose, because otherwise
 // LanSuite might not work properly anymore.
